requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
lxml==5.1.0
pyasn1==0.5.1
pysmb==1.2.9.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.1.0
webdavclient3==3.14.6
